requirements.txt
torch>=2.0.0
pytorch-ignite>=0.5.0
pyyaml>=6.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.14.0
